fix monetary precision and validation gaps in checkout flow
- Replace double with BigDecimal for unitPrice (OrderItem) and totalPrice
  (FulfillmentOrder, CheckoutResponse, FulfillmentOrderResponse) so
  precision is maintained end-to-end without doubleValue() conversions
- Add @Valid cascade on CheckoutRequest.items and @NotBlank/@Min(1)
  constraints on OrderItem so invalid items are rejected at the HTTP
  layer with a clean 400 instead of reaching the service
- Handle MethodArgumentNotValidException in GlobalExceptionHandler so
  @Valid failures return {"error":"..."} consistent with other errors
- Fix test item bodies in address-null tests to use complete OrderItem
  data, ensuring the 400 is triggered by the address constraint and not
  a future item constraint
- Add verify(checkoutService).checkout(any()) to exception-handling
  tests and replace deprecated isUnprocessableEntity() with is(422)

+++ backend/src/main/java/org/example/backend/order/dto/CheckoutResponse.java
@@ -2,10 +2,12 @@ package org.example.backend.order.dto;
 
 import org.example.backend.order.model.FulfillmentOrderStatus;
 
+import java.math.BigDecimal;
+
 public record CheckoutResponse(
         String orderNumber,
         String invoiceNumber,
         FulfillmentOrderStatus status,
-        double totalPrice
+        BigDecimal totalPrice
 ) {
 }
